WPT-Project1-Folder(IntelligentInventoryOptimizationAndStockoutPrediction)/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
flask-caching==2.1.0
gunicorn==21.2.0
